validation update for cv

+++ src/controllers/cv.controller.ts
@@ -5,6 +5,7 @@ import { pick } from "@/common/utils/pick.utils";
 import { cvService } from "@/services/cv.service";
 import { StatusCodes } from "http-status-codes";
 
+import { UserRole } from "@/types";
 import type { Request, Response } from "express";
 
 class CVController {
@@ -12,12 +13,18 @@ class CVController {
     if (!req.user) {
       throw new ApiError(StatusCodes.NOT_FOUND, "User not found");
     }
+    if (!req.role || req.role !== UserRole.TALENT) {
+      throw new ApiError(StatusCodes.UNAUTHORIZED, "Unauthorized access");
+    }
     const userId = req.user.id;
     const serviceResponse = await cvService.create(req.body, userId);
     return handleServiceResponse(serviceResponse, res);
   });
 
-  public getCv = catchAsync(async (req: Request, res: Response) => {
+  public getCvById = catchAsync(async (req: Request, res: Response) => {
+    if (!req.user) {
+      throw new ApiError(StatusCodes.NOT_FOUND, "User not found");
+    }
     const { id } = req.params;
     const serviceResponse = await cvService.getCv(id);
     return handleServiceResponse(serviceResponse, res);
@@ -33,17 +40,34 @@ class CVController {
   });
 
   public getAllCvs = catchAsync(async (req: Request, res: Response) => {
+    if (!req.user) {
+      throw new ApiError(StatusCodes.NOT_FOUND, "User not found");
+    }
     const serviceResponse = await cvService.getAllCvs();
     return handleServiceResponse(serviceResponse, res);
   });
 
   public deleteCv = catchAsync(async (req: Request, res: Response) => {
+    if (!req.user) {
+      throw new ApiError(StatusCodes.NOT_FOUND, "User not found");
+    }
+    if (!req.role || req.role !== UserRole.TALENT) {
+      throw new ApiError(StatusCodes.UNAUTHORIZED, "Unauthorized access");
+    }
+
     const { id } = req.params;
     const serviceResponse = await cvService.deleteCv(id);
     return handleServiceResponse(serviceResponse, res);
   });
 
   public updateCv = catchAsync(async (req: Request, res: Response) => {
+    if (!req.user) {
+      throw new ApiError(StatusCodes.NOT_FOUND, "User not found");
+    }
+    if (!req.role || req.role !== UserRole.TALENT) {
+      throw new ApiError(StatusCodes.UNAUTHORIZED, "Unauthorized access");
+    }
+
     const { id } = req.params;
     const serviceResponse = await cvService.updateCv(id, req.body);
     return handleServiceResponse(serviceResponse, res);
